use double for records discord too

+++ src/game/server/gamecontext.h
@@ -637,7 +637,7 @@ public:
 	//+KZ
 	IHttp *m_pHttp;
 	void SendDiscordChatMessage(int ClientID, const char* msg);
-	void SendDiscordRecordMessage(int ClientID, float Time, float PrevTime);
+	void SendDiscordRecordMessage(int ClientID, double Time, double PrevTime);
 
 	void SendGameMsg(int GameMsgId, int ClientId) const;
 	void SendGameMsg(int GameMsgId, int ParaI1, int ClientId) const;
